+++ src/gdsii.cpp
@@ -42,6 +42,7 @@ double gdsii_real_to_double(uint64_t real) {
 uint32_t gdsii_read_record(FILE* in, uint8_t* buffer) {
     uint64_t read_length = fread(buffer, 1, 4, in);
     if (read_length < 4) {
+        DEBUG_PRINT("Read bytes (expected 4): %" PRIu64 "\n", read_length);
         if (feof(in) != 0)
             fputs("[GDSTK] Unable to read input file. End of file reached unexpectedly.\n", stderr);
         else
@@ -50,10 +51,17 @@ uint32_t gdsii_read_record(FILE* in, uint8_t* buffer) {
     }
     big_endian_swap16((uint16_t*)buffer, 1);  // second word is interpreted byte-wise (no swaping);
     const uint32_t record_length = *((uint16_t*)buffer);
-    if (record_length < 4) return 0;
-    if (record_length == 4) return record_length;
+    if (record_length < 4) {
+        DEBUG_PRINT("Record length should be at beast 4. Found %" PRIu32 "\n", record_length);
+        fputs("[GDSTK] Corrupted GDSII file.\n", stderr);
+        return 0;
+    } else if (record_length == 4) {
+        return record_length;
+    }
     read_length = fread(buffer + 4, 1, record_length - 4, in);
     if (read_length < record_length - 4) {
+        DEBUG_PRINT("Read bytes (expected %" PRIu32 "): %" PRIu64 "\n", record_length - 4,
+                    read_length);
         if (feof(in) != 0)
             fputs("[GDSTK] Unable to read input file. End of file reached unexpectedly.\n", stderr);
         else